requirements.txt
pytest==8.3.4
requests==2.32.3
PyYAML==6.0.2
pytest-html==4.1.1
